+++ NEWS.md
@@ -3,6 +3,6 @@
 NEW FEATURES
 
 * Converted `dasper` into a Bioconductor-friendly format using [biocthis](https://lcolladotor.github.io/biocthis/)
-* Added `junction_load()`, which loads raw junction data outputted from RNA-sequencing alignmetn into an `RangedSummarizedExperiment` object
+* Added `junction_load()`, which loads raw junction data outputted from RNA-sequencing alignment into an `RangedSummarizedExperiment` object. Includes an option to allow download of user-specified control junctions.